Fixes #33 - Zoho's documentation is wrong.
When trying to void a credit note, the Books API was crashing
saying that it needed a JSONString param. At a guess, there's
some checks in their backend for a POST, and it must contain
JSONData before it gets processed.

THIS IS WHY REST EXISTS, ZOHO. YOU SHOULD BE USING A PATCH, NOT
A POST CALL.